permission_filter: fix empty query_filters bug, test and refactor
There was a problem (a form of Liskov violation) with permission_filter where
None would be returned if query_filters returned []. This would break the search.
Now, that issue is solved, permission_filter is tested, and refactored a little.

A better refactoring IMO would place the permission_filter code in Permission.query_filters
directly and remove the need for search to call this dangling function.